backend_project/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.5
dotenv==0.9.9
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
FreeSimpleGUI==5.2.0.post1
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.3
